Variable Selection and Profile Analysis | Information Value and Weight of Evidence
From the given set of variables, we can identify (in a naive manner) which of the predictors(X) will affect the Credit Score(BAD) in which manner.
An intuitive list of such parameters would be:
DOB
We can calculate the AGE of a person, given their date of Birth. For the rest of this analysis, we will be considering AGE = 99-DOB for our analysis. (AGE = 0 indicates no data given)
Prior knowledge indicates that good credit scores are given to people in their mid-lives (25-45). They have a high propensity to earn and to spend. People aged <25 and >45 are usually considered as “non-earning” and have low credit ratings.
NKID
Number of children inversely affects credit rating. Having children is an expensive affair involving heavy expenditure. One could argue that being a parent makes one a cautious spender, so this assumption needs validation.
DEP
Increased number of Dependents affects credit rating inversely. Credit score goes down as dependents increase.
PHON
In 1999, having a phone was considered a rarity so owning one should indicate good credit.
SINC and DAINC
This gives the total income in the Household and a greater value gives a better indication of good credit.
AES
Type of employment (especially ones that indicate whether employed or unemployed) is a good indicator of credit.
RES
Homeowners tend to have good credit, provided their mortgage payments aren’t very high (DMORT). This term most likely involves interaction with DMORT.
DHVAL
Value of Home Asset, along with a low DMORT would indicate good credit. These terms may interact and we need to consider them in our Model later on.
DOUTM/DOUTL/DOUTCC/DOUTHP
These indicate the Outgoing expenditure for a person and inversely affect credit scores.
BAD
This is the response variable we need to predict. BAD = 1 indicates a bad credit score and BAD = 0 indicates a good credit score. (Technically, NOT BAD credit score)
Overall, each of the above variables seem to have an effect on Credit Scores(BAD) in one way or another. In order to determine which the important variables are, we can individually check for each of them as shown in Appendix 1.