Optimal Cut-off Calculations
One issue noted in the dataset is how unbalanced the data appears to be. There are more GOOD credit scores in the data compared to BAD credit scores. This has to be kept in mind when selecting a cut-off probability value for the Logistic Regression Model. For the Model built in the previous step